+++ backend/core/types/src/backend-swagger.ts
@@ -800,6 +800,48 @@ export class UserService {
   /**
    * Resend confirmation
    */
+  resendPartnerConfirmation(
+    params: {
+      /** requestBody */
+      body?: Email
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<Status> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/user/resend-partner-confirmation"
+
+      const configs: IRequestConfig = getConfigs("post", "application/json", url, options)
+
+      let data = params.body
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Verifies token is valid
+   */
+  isUserConfirmationTokenValid(
+    params: {
+      /** requestBody */
+      body?: Confirm
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<boolean> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/user/is-confirmation-token-valid"
+
+      const configs: IRequestConfig = getConfigs("post", "application/json", url, options)
+
+      let data = params.body
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Resend confirmation
+   */
   resendConfirmation(
     params: {
       /** requestBody */
@@ -3875,6 +3917,11 @@ export interface PaginatedAssets {
   meta: PaginationMeta
 }
 
+export interface UserErrorExtraModel {
+  /**  */
+  userErrorMessages: EnumUserErrorExtraModelUserErrorMessages
+}
+
 export interface Login {
   /**  */
   email: string
@@ -4023,6 +4070,9 @@ export interface User {
 
   /**  */
   agreedToTermsOfService: boolean
+
+  /**  */
+  hitConfirmationURL?: Date
 }
 
 export interface UserCreate {
@@ -4067,6 +4117,9 @@ export interface UserCreate {
 
   /**  */
   phoneNumberVerified?: boolean
+
+  /**  */
+  hitConfirmationURL?: Date
 }
 
 export interface UserBasic {
@@ -4132,6 +4185,9 @@ export interface UserBasic {
 
   /**  */
   agreedToTermsOfService: boolean
+
+  /**  */
+  hitConfirmationURL?: Date
 }
 
 export interface Email {
@@ -4236,6 +4292,9 @@ export interface UserUpdate {
 
   /**  */
   agreedToTermsOfService: boolean
+
+  /**  */
+  hitConfirmationURL?: Date
 }
 
 export interface UserFilterParams {
@@ -4301,6 +4360,9 @@ export interface UserInvite {
 
   /**  */
   phoneNumberVerified?: boolean
+
+  /**  */
+  hitConfirmationURL?: Date
 }
 
 export interface UserProfileUpdate {
@@ -6464,6 +6526,16 @@ export enum EnumApplicationsApiExtraModelOrderBy {
 export enum EnumApplicationsApiExtraModelOrder {
   "ASC" = "ASC",
   "DESC" = "DESC",
+}
+export enum EnumUserErrorExtraModelUserErrorMessages {
+  "accountConfirmed" = "accountConfirmed",
+  "accountNotConfirmed" = "accountNotConfirmed",
+  "errorSaving" = "errorSaving",
+  "emailNotFound" = "emailNotFound",
+  "tokenExpired" = "tokenExpired",
+  "tokenMissing" = "tokenMissing",
+  "emailInUse" = "emailInUse",
+  "passwordOutdated" = "passwordOutdated",
 }
 export enum EnumLoginMfaType {
   "sms" = "sms",