+++ ui-components/src/helpers/useMutate.ts
@@ -1,12 +1,20 @@
 import { useState } from "react"
 
+export type UseMutateOptions = {
+  onSuccess?: () => void
+  onError?: (err: any) => void
+}
+
 const useMutate = <UseMutateResponse>() => {
   const [data, setData] = useState<UseMutateResponse | undefined>(undefined)
   const [isSuccess, setSuccess] = useState(false)
   const [isLoading, setLoading] = useState(false)
   const [isError, setError] = useState<unknown>(null)
 
-  const mutate = async (mutateFn: (args?: unknown) => Promise<UseMutateResponse>) => {
+  const mutate = async (
+    mutateFn: (args?: unknown) => Promise<UseMutateResponse>,
+    options?: UseMutateOptions
+  ) => {
     let response: UseMutateResponse | undefined = undefined
 
     try {
@@ -14,11 +22,13 @@ const useMutate = <UseMutateResponse>() => {
       response = await mutateFn()
       setData(response)
       setSuccess(true)
+      options?.onSuccess?.()
       setLoading(false)
     } catch (err) {
       setSuccess(false)
       setLoading(false)
       setError(err)
+      options?.onError?.(err)
     }
 
     return response